Fix up nudge rendering.
