--- apps/full-site-editing/full-site-editing-plugin/donations/donations-tabs.js
@@ -1,57 +1,0 @@
-/**
- * External dependencies
- */
-import classNames from 'classnames';
-
-/**
- * WordPress dependencies
- */
-import { useState } from '@wordpress/element';
-import { __ } from '@wordpress/i18n';
-import { Button } from '@wordpress/components';
-
-/**
- * Internal dependencies
- */
-import StripeNudge from './stripe-nudge';
-
-const DonationsTabs = ( { stripeConnectUrl } ) => {
-	const [ activeTab, setActiveTab ] = useState( 'one-time' );
-
-	const isActive = ( button ) => ( activeTab === button ? 'active' : null );
-
-	return (
-		<div className="donations__container">
-			{ stripeConnectUrl && <StripeNudge stripeConnectUrl={ stripeConnectUrl } /> }
-			<Button className={ isActive( 'one-time' ) } onClick={ () => setActiveTab( 'one-time' ) }>
-				{ __( 'One-Time', 'full-site-editing' ) }
-			</Button>
-			<Button className={ isActive( 'monthly' ) } onClick={ () => setActiveTab( 'monthly' ) }>
-				{ __( 'Monthly', 'full-site-editing' ) }
-			</Button>
-			<Button className={ isActive( 'annually' ) } onClick={ () => setActiveTab( 'annually' ) }>
-				{ __( 'Annually', 'full-site-editing' ) }
-			</Button>
-			<div
-				id="donations__tab-one-time"
-				className={ classNames( 'donations__tab', { active: isActive( 'one-time' ) } ) }
-			>
-				{ __( 'One time', 'full-site-editing' ) }
-			</div>
-			<div
-				id="donations__tab-monthly"
-				className={ classNames( 'donations__tab', { active: isActive( 'monthly' ) } ) }
-			>
-				{ __( 'Monthly', 'full-site-editing' ) }
-			</div>
-			<div
-				id="donations__tab-annually"
-				className={ classNames( 'donations__tab', { active: isActive( 'annually' ) } ) }
-			>
-				{ __( 'Annually', 'full-site-editing' ) }
-			</div>
-		</div>
-	);
-};
-
-export default DonationsTabs;
