+++ apps/full-site-editing/full-site-editing-plugin/donations/edit.js
@@ -7,7 +7,7 @@ import { __ } from '@wordpress/i18n';
 /**
  * Internal dependencies
  */
-import DonationsTabs from './donations-tabs';
+import Donations from './donations';
 import LoadingError from './loading-error';
 import LoadingStatus from './loading-status';
 import UpgradePlan from './upgrade-plan';
@@ -54,7 +54,7 @@ const Edit = () => {
 		return <UpgradePlan upgradeUrl={ upgradeUrl } />;
 	}
 
-	return <DonationsTabs stripeConnectUrl={ stripeConnectUrl } />;
+	return <Donations stripeConnectUrl={ stripeConnectUrl } />;
 };
 
 export default Edit;
